app/requirements/base.txt
asgiref==3.5.2
attrs==19.3.0
autobahn>=20.3.1
Automat==20.2.0
auto_dataclass==0.1.6
cachetools==5.3.0
certifi==2019.11.28
channels==3.0.5
chardet==3.0.4
constantly==15.1.0
cryptography>=2.7,<41.0.5
redis==4.3.4
hiredis==2.0.0
daphne>=2.4,<4.0
Django==3.2.16
django-cors-headers==3.13.0
django-eventstream==3.1.0
django-extensions==3.2.3
django-grip==3.0.0
django-jsonfield==1.4.1
django-otp==1.2.4
django-ratelimit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-redis==5.2.0
djangorestframework-simplejwt==5.3.0
django-debug-toolbar==3.8.1
djoser==2.0.3
django-coverage-plugin==2.0.3  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.5.2  # https://github.com/pytest-dev/pytest-django
drf-spectacular==0.27.2
gripcontrol==4.0.0
importlib-metadata==4.13.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
multiprocess==0.70.14
nest-asyncio==1.6.0
networkx==2.6.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
platformdirs==2.6.2
pox==0.3.2
ppft==1.7.6.6
premailer==3.10.0
prompt-toolkit==3.0.43
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyHamcrest==2.0.2
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.10
PyJWT==1.7.1
pytz==2019.3
pyOpenSSL>=22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pystream-protobuf==1.6.4
qrcode==7.4.2
requests==2.23.0
rsa==4.9
s3transfer==0.6.0
schema==0.7.5
secure-smtplib==0.1.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.4.3
Twisted==20.3.0
txaio==20.1.1
urllib3>=1.21.1,<1.26.5
Werkzeug==0.16.1
zope.interface==5.0.0
faust==1.10.4
bson==0.5.10
h11==0.14.0 
uvicorn==0.22.0
websockets==11.0.3